% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/abund_comp.R
\name{abund_comp}
\alias{abund_comp}
\title{abund_comp.R}
\usage{
abund_comp(
  SurveyNamein,
  DirNameFilein,
  imagedirin,
  casesin,
  hakeflagin,
  yearin,
  ...
)
}
\arguments{
\item{casesin}{- is the set of cases to be compared}

\item{hakeflagin}{adult hake (hakeflag=2), not filtering for hake (hakeflag=-1), age-0 hake (hakeflag=0), age-1 hake only (hakeflag=1), all hake (hakeflag=3), age-1+ (hakeflag=4). Passing this although assuming this will be hakeflag=1}

\item{SurveyName}{- name of survey}

\item{DirNameFile}{excel file - contains paths and locations of files. Uses
Export_dir as the directory that the exported files are in.}

\item{imagedir}{- the directory for the images files to be placed in}

\item{year}{is the year that is being investigated}
}
\description{
Function to bring in and compare age2+ abundances
}
\examples{
 \dontrun{
aB2<-abund_comp(SurveyName,DirNameFile,imagedir, casesin=cases, hakeflagin,yearin=year)
}
}
