% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/basic_BM_comps.R
\name{basic_BM_comps}
\alias{basic_BM_comps}
\title{basic_BM_comps.R}
\usage{
basic_BM_comps(
  SurveyNamein,
  DirNameFilein,
  imagedirin,
  case1in,
  case2in,
  hakeflagin = 2,
  ...
)
}
\arguments{
\item{SurveyName}{- name of survey}

\item{DirNameFile}{excel file - contains paths and locations of files. Uses
Export_dir as the directory that the exported files are in.}

\item{sheet}{is an optional argument to denote which sheet of the DirNameFile
want to use. This defaults to the first sheet if not set.}

\item{hakeflag:}{adult hake age2+ (hakeflag=2), not filtering for hake (hakeflag=-1),
age-0 hake (hakeflag=0), age-1 hake only (hakeflag=1), all hake (hakeflag=3), age-1+ (hakeflag=4).   Defaults to age2+}
}
\description{
Function to compare two biomass estimates generated by Matlab
}
\examples{
 \dontrun{
T<-basic_BM_comps(data,data2)
}
}
