% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/basic_BM_comps_mult.R
\name{basic_BM_comps_mult}
\alias{basic_BM_comps_mult}
\title{basic_BM_comps_mult.R}
\usage{
basic_BM_comps_mult(
  SurveyNamein,
  DirNameFilein,
  imagedirin,
  cases,
  hakeflagin,
  yearin,
  ...
)
}
\arguments{
\item{cases}{- is the set of cases to be compared}

\item{hakeflagin}{adult hake (hakeflag=2), not filtering for hake (hakeflag=-1), age-0 hake (hakeflag=0), age-1 hake only (hakeflag=1), all hake (hakeflag=3), age-1+ (hakeflag=4). Passing this although assuming this will be hakeflag=1}

\item{yearin}{is the year that is being investigated}

\item{SurveyName}{- name of survey}

\item{DirNameFile}{excel file - contains paths and locations of files. Uses
Export_dir as the directory that the exported files are in.}

\item{imagedir}{- the directory for the images files to be placed in}
}
\description{
Function to compare two biomass estimates generated by Matlab
}
\examples{
 \dontrun{
T<-basic_BM_comps_mult(SurveyName,DirNameFile,imagedir,cases,hakeflagin, yearin)
}
}
