% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/basic_comps.R
\name{basic_comps}
\alias{basic_comps}
\title{basic_comps.R}
\usage{
basic_comps(X, Y, compcols, imagedirin, case1in, case2in, ...)
}
\arguments{
\item{X}{data frame}

\item{Y}{second data frame}

\item{compcols}{the names of columns to be compared for certain comparisons}
}
\description{
Function to compare two dataframes of Echoview file exports.  
The output is a list along with plots.  The first item in the list are Boolean values that reflect
the answers to the following questions:
The second item in the list are the row and columns of differences between the 
two dataframes.
}
\examples{
 \dontrun{
T<-basiccomps(data,data2)
}
}
