% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/basic_comps_mult.R
\name{basic_comps_mult}
\alias{basic_comps_mult}
\title{basic_comps_mult.R}
\usage{
basic_comps_mult(datalist, compcols, imagedirin, cases, hakeflag, ...)
}
\arguments{
\item{datalist}{- list of dataframes}

\item{compcols}{- the names of columns to be compared for certain comparisons}

\item{imagedirin}{- the directory to save the images/tables in}

\item{cases}{- a list of text strings with the case names for each datafram}

\item{hakeflag}{- adult hake (hakeflag=2), not filtering for hake (hakeflag=-1), 
age-0 hake (hakeflag=0), age-1 hake only (hakeflag=1), all hake (hakeflag=3), age-1+ (hakeflag=4).}
}
\description{
Function to compare several dataframes of Echoview file exports.  
The output is a list along with plots.  The first item in the list are Boolean values that reflect
the answers to the following questions:
The second item in the list are the row and columns of differences between the 
two dataframes.
}
\examples{
 \dontrun{
T<-basic_comps_mult(datalist,compcols,imagedirin,cases,,hakeflag[k])
}
}
