% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compare_NASC_strata.R
\name{compare_NASC_strata}
\alias{compare_NASC_strata}
\title{compare_NASC_strata:  Read in report(s) generated by Matlab biomass estimation/kriging process}
\usage{
compare_NASC_strata(SurveyName, DirNameFile, year, cases = NULL, ...)
}
\arguments{
\item{SurveyName}{- name of survey}

\item{DirNameFile}{excel file - contains paths and locations of files. Uses
Export_dir as the directory that the exported files are in.}

\item{cases}{(cases to be compared - for right now, just compare 2)}
}
\description{
Read in NASC tables for age-1+ full agemethods from two cases and compare
}
\examples{
temp<-compare_NASC_strata(SurveyName, DirNameFile,year,cases)

}
