% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getTnumv1.R
\name{getTnumv1}
\alias{getTnumv1}
\title{getTnumv2.R}
\usage{
getTnumv1(A)
}
\arguments{
\item{A:}{string that is the name of the EV file}
}
\description{
Function to pull out the transect number from the standard exported
csv convention of V###-S#####-X#-F38-T##-Z0- ( ).csv where the number after the V is the
vessel, after the S is the survey number, after the X is the transducer, after the F
is the frequency (generally 38 kHz), and after the T is the transect number
}
\examples{
 \dontrun{
Tr<-getTnumv1("V499-S200952-X1-F38-T98-Z0- (analysis).csv")
}
}
