% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getTnumv2.R
\name{getTnumv2}
\alias{getTnumv2}
\title{getTnumv2.R}
\usage{
getTnumv2(A)
}
\arguments{
\item{A:}{string that is the name of the EV file}
}
\description{
Function to pull out the transect number form the name of an EV by using the 
 number after the 'x'
}
\examples{
 \dontrun{
Tr<-getTnumv2("x040hake69_raw.EV")
}
}
