% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getTnumv3.R
\name{getTnumv3}
\alias{getTnumv3}
\title{getTnumv3.R}
\usage{
getTnumv3(A)
}
\arguments{
\item{A:}{string that is the name of the EV file}
}
\description{
Function to pull out the transect number form the name of an EV by using the
 number after the last '/' in the filename.  Run if there is no "x" in the filename
}
\examples{
 \dontrun{
Tr<-getTnumv3("C:\\\\rthomas\\\\Projects\\\\EV_compilation\\\\Local\\\\2003\\\\EV files\\\\age1\\\\original\\\\CAN/03gwf.ev")
}
}
