% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_csv_addflnm.R
\name{read_csv_addflnm}
\alias{read_csv_addflnm}
\title{read_csv_addflnm: Function to read in .csv files and append filename into new
column}
\usage{
read_csv_addflnm(flnm)
}
\arguments{
\item{flnm}{- filename (full path) of csv file to read in}
}
\description{
read_csv_addflnm is a function to read in .csv files (e.g. Echoview exports)
as a table and append the filename as another column in the table.  Is based
off read_csv (from package readr)
}
\examples{
\dontrun{
}
}
