% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_EVdir_exports.R
\name{read_EVdir_exports}
\alias{read_EVdir_exports}
\title{read_EV_exports:  Read in a directory of .csv files exported from Echoview}
\usage{
read_EVdir_exports(
  SurveyName,
  DirNameFile,
  database = 1,
  exportloc = 1,
  sheet = NULL,
  hakeflag = 2,
  ...
)
}
\arguments{
\item{SurveyName}{- name of survey}

\item{DirNameFile}{excel file - contains paths and locations of files. Uses
Export_dir as the directory that the exported files are in.}

\item{database}{is an optional argument for format of the .csv files to.  Default is database (1).  Spreadsheet =0}

\item{exportloc}{is an optional argument for using the Export_Dir or the Region_Export_dir column
entry in the DirNameFile spreadsheet.  1 is Export_Dir, and 0 is Region_Export_dir}

\item{sheet}{is an optional argument to denote which sheet of the DirNameFile
want to use. This defaults to the first sheet if not set.}

\item{hakeflag}{sets whether specific "hake"-type classes are kept.  hakeflag=0 keeps just age-0 hake,
hakeflag=1 keeps just age-1 hake, hakeflag=2 keeps just regular age2+ hake, hakeflag=3 keeps everything that could
possibly be hake.  Any other numbers (including -1) will keep everything regardless of whether it is hake.
The default is hakeflag=2, keeping just adult hake.}
}
\description{
Read in a directory of .csv files (can be of several variables)
exported from Echoview.   The function preserves the cells by depth.
Uses packages readxl, readr, purrr, and dplyr.
}
\examples{
temp<-read_EV_exports(SurveyName, DirNameFile,database=1,exportloc=1)

}
