% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_proc_reports_abund.R
\name{read_proc_reports_abund}
\alias{read_proc_reports_abund}
\title{read_proc_reports_abund:  Read in report(s) generated by Matlab biomass estimation/kriging process}
\usage{
read_proc_reports_abund(
  SurveyName,
  DirNameFile,
  sheet = NULL,
  hakeflag = 2,
  ...
)
}
\arguments{
\item{SurveyName}{- name of survey}

\item{DirNameFile}{excel file - contains paths and locations of files. Uses
Export_dir as the directory that the exported files are in.}

\item{sheet}{is an optional argument to denote which sheet of the DirNameFile
want to use. This defaults to the first sheet if not set.}

\item{hakeflag:}{adult hake (hakeflag=2), not filtering for hake (hakeflag=-1),
age-0 hake (hakeflag=0), age-1 hake only(hakeflag=1), all hake (hakeflag=3), age-1+ (hakeflag=4).
set to 1 if want age1+ numbers, set to 0 (or other number) if want age2+.  Defaults to age2+}
}
\description{
Read in report(s) from Matlab abundance estimation from a directory
and calculate the total survey abundance estimate and abundance by age. Uses readxl
}
\examples{
temp<-read_proc_reports_abund(SurveyName, DirNameFile,sheet, hakeflag)

}
