% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_proc_reports_abund_cross.R
\name{read_proc_reports_abund_cross}
\alias{read_proc_reports_abund_cross}
\title{read_proc_reports_abund_cross:  Read in report(s) generated by Matlab biomass estimation/kriging process}
\usage{
read_proc_reports_abund_cross(
  SurveyName,
  DirNameFile,
  sheet = NULL,
  hakeflag = 2,
  method = 1,
  ...
)
}
\arguments{
\item{SurveyName}{- name of survey}

\item{DirNameFile}{excel file - contains paths and locations of files. Uses
Export_dir as the directory that the exported files are in.}

\item{sheet}{is an optional argument to denote which sheet of the DirNameFile
want to use. This defaults to the first sheet if not set.}

\item{hakeflag:}{adult hake (hakeflag=2), not filtering for hake (hakeflag=-1),
age-0 hake (hakeflag=0), age-1 hake only(hakeflag=1), all hake (hakeflag=3), age-1+ (hakeflag=4).
set to 1 if want age1+ numbers, set to 0 (or other number) if want age2+.  Defaults to age2+}

\item{method:}{Can read in age-2+ biomass for either age-1+ method () or age-2+ method.
Cannot read in age-1 biomass for age-2+ method.  Method = 1 is for age-1+ method, method = 2
is for the age-2+ method.}
}
\description{
Read in report(s) from Matlab kriged abundance estimation from a directory
and calculate the total survey abundance estimate and abundance by age. Uses readxl
Is able to read in with separate method (age-1 or age-2+ method) and hakeflag.
}
\examples{
\dontrun{
temp<-read_proc_reports_cross(SurveyName, DirNameFile,sheet,hakeflag, method)
}

}
