% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Read_Report_Summaries.R
\name{Read_Report_Summaries}
\alias{Read_Report_Summaries}
\title{Read_Report_Summaries}
\usage{
Read_Report_Summaries(repcases, year, reportpath, hakeflag = 2)
}
\arguments{
\item{repcases:}{a list of cases (will correspond to subfolder names) of reports to read in.}

\item{year:}{the year (used as a subfolder)}

\item{reportpath:}{the base path that the script will look for the (year then) case subfolders in}

\item{hakeflag:}{adult hake (hakeflag=2), not filtering for hake (hakeflag=-1),
age-0 hake (hakeflag=0), age-1 hake only(hakeflag=1), all hake (hakeflag=3), age-1+ (hakeflag=4).
set to 1 if want age1+ numbers, set to 0 (or other number) if want age2+.  Defaults to age2+}
}
\description{
helper function to read in biomass at age reports from EchoPro from different cases
as well as with different methods (age1+ (full age) and age2+ methods).  Right now just tested for age-2+ results
for both methods
}
\examples{
 \dontrun{
Blist<-Read_Report_Summaries(repcases, year, reportpath, hakeflag)
}
}
