% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_spat_unkriged_reports.R
\name{read_spat_unkriged_reports}
\alias{read_spat_unkriged_reports}
\title{read_spat_unkriged_reports:  Read in report(s) generated by Matlab biomass estimation/kriging process}
\usage{
read_spat_unkriged_reports(
  SurveyName,
  DirNameFile,
  sheet = NULL,
  hakeflag = 2,
  ...
)
}
\arguments{
\item{SurveyName}{- name of survey}

\item{DirNameFile}{excel file - contains paths and locations of files. Uses
Export_dir as the directory that the exported files are in.}

\item{sheet}{is an optional argument to denote which sheet of the DirNameFile
want to use. This defaults to the first sheet if not set.}

\item{hakeflag.}{adult hake (hakeflag=2), not filtering for hake (hakeflag=-1),
age-0 hake (hakeflag=0), age-1 hake only (hakeflag=1), all hake (hakeflag=3), age-1+ (hakeflag=4).}
}
\description{
Read in report(s) from Matlab biomass estimation from a directory
and calculate the total survey biomass estimate and biomass by age. Uses readxl
}
\examples{
temp<-read_spat_unkriged_reports(SurveyName, DirNameFile,sheet,hakeflag=2)

}
