% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Read_Spat_Unkriged_Reports_cross.R
\name{Read_Spat_Unkriged_Reports_cross}
\alias{Read_Spat_Unkriged_Reports_cross}
\title{Read_Spat_Unkriged_Reports_cross}
\usage{
Read_Spat_Unkriged_Reports_cross(
  repcases,
  year,
  projpathbase,
  estimflag = 4,
  hakeflag = 2
)
}
\arguments{
\item{repcases:}{a list of cases (will correspond to subfolder names) of reports to read in.}

\item{year:}{the year (used as a subfolder)}

\item{projpathbase:}{the base path that the project is in}

\item{estimflag:}{estimation method (age-2+ or age-1+ method).  Age-1+ method is 4,
age-2+ method is 2 (or any number other than 4)}

\item{hakeflag:}{adult hake (hakeflag=2), not filtering for hake (hakeflag=-1),
age-0 hake (hakeflag=0), age-1 hake only(hakeflag=1), all hake (hakeflag=3), age-1+ (hakeflag=4).
set to 1 if want age1+ numbers, set to 0 (or other number) if want age2+.  Defaults to age2+}
}
\description{
helper function to read in spatial biomass at age reports from EchoPro. This version
is different from Read_spat_unkriged_Reports because it allows you to read in an age-2+ estimate from
an age-1+ estimation process.  Will keep just the stratum name and total biomass (no at-age data)
}
\examples{
 \dontrun{
Blist<-Read_Spat_Unkriged_Reports_cross(repcases, year, projpathbase,estimflag,hakeflag)
}
}
