% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/repath_EVdir.R
\name{repath_EVdir}
\alias{repath_EVdir}
\title{repath_EVdir.R}
\usage{
repath_EVdir(SurveyName, DirNameFile, ni = 1, ...)
}
\arguments{
\item{SurveyName}{name of survey as in excel file}

\item{DirNameFile}{excel file name of paths}

\item{ni}{is an optional argument for which file to start with (default ni=1 is the first file).

Excel file has the following columns (with header names)
Cal_File: calibration file name
Base_Path:
Orig_EV_Dir:
Raw_Dir
EK80_Cal_File
EK80_Raw_Dir}
}
\description{
Function to take a directory of EV files and reset the path.
This works for EV files that have just an EK80 fileset, or that have the EK60 file set
as the first fileset, and the EK80 as the second fileset.
May want to eventually convert to calling "fileset1" and "fileset2" instead of EK60 and EK80.
Written by RT 2022.
Requires an excel file input hat says what the paths are
Each row is a survey (survey_name), and each column is a path or file name
}
\examples{
 \dontrun{
DirNameFile="C:/rthomas/R/Rcode/EK60_EK80 conversion/EK60_Ek80_conv_updated/Directory Structure EK60 EK80 conversion updated.xlsx"
SurveyName='2019_US'
repath_EVdir(SurveyName,DirNameFile)
}
}
