% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/repath_EVdir_flex2.R
\name{repath_EVdir_flex2}
\alias{repath_EVdir_flex2}
\title{repath_EVdir_flex2.R}
\usage{
repath_EVdir_flex2(SurveyName, DirNameFile, ni = 1, sheet, ...)
}
\arguments{
\item{SurveyName}{name of survey as in excel file}

\item{DirNameFile}{excel file name of paths
Excel file has the following columns (with header names)
Cal_File: calibration file name
Base_Path:
Orig_EV_Dir:
Raw_Dir
EK5_Dir_Flag
EK5_Raw_Dir}

\item{ni}{is an optional argument for which file to start with (default ni=1 is the first file).}

\item{sheet}{is an optional argument to denote which sheet of the DirNameFile}
}
\description{
Function to take a directory of EV files and reset the path.
This works for EV files that have an ek5 sub-directory too.
The ek5 subdirectory must be named 'ek5 EV'.  The EK5_Dir_Flag must be
set appropriately in the excel file. This version saves to the directory in
Post_EV_Dir
Written by RT 2022, modified 2024.
Requires an excel file input that says what the paths are
Each row is a survey (survey_name), and each column is a path or file name
}
\examples{
 \dontrun{
DirNameFile="C:/rthomas/R/Rcode/EK60_EK80 conversion/EK60_Ek80_conv_updated/Directory Structure EK60 EK80 conversion updated.xlsx"
SurveyName='2019_US'
repath_EVdir_flex(SurveyName,DirNameFile, ni, sheet)
}
}
