% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Check_EV_Properties.R
\name{Check_EV_Properties}
\alias{Check_EV_Properties}
\title{Check_EV_Properties.R}
\usage{
Check_EV_Properties(SurveyName, DirNameFile, variables, sheet = case)
}
\arguments{
\item{SurveyName}{name of survey as in excel file}

\item{DirNameFile}{excel file name of paths}

\item{variables}{- variables to export}

\item{sheet}{is an optional argument to denote which sheet of the DirNameFile
want to use. This defaults to the first sheet if not set.}

\item{database}{is an optional argument for export format.  Default is database (1).  Spreadsheet =0}
}
\description{
Uses COM objects to run Echoview and get transducer properties.  These properties are 
saved to a .csv file in the Data_Compile_dir of the DirNameFile excel spreadsheet. The name defaults
to TransducerConfigCheck(SurveyName).csv
}
\examples{
 \dontrun{
Check_EV_Properties(SurveyName1in, DirNameFile,variables,sheet=case)
}
}
