% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/export_integration_EVdir.R
\name{export_integration_EV}
\alias{export_integration_EV}
\title{export_integration_EVdir}
\usage{
export_integration_EV(
  Survey_Name,
  DirNameFile,
  variables,
  paramnamefile = NA,
  ni = 1,
  gdepth = 10,
  database = 1,
  sheet,
  ...
)
}
\arguments{
\item{DirNameFile}{excel file name of paths}

\item{variables}{- variables to export (will name subdirectories with variables)}

\item{paramnamefile}{optional excel file name of parameters to export or not export.
The column ToExport is parameters to export.  The column NoExport is parameters to not export}

\item{ni}{is an optional argument for which file to start with (default ni=1 is the first file).
This defaults to 1 (all EV files)}

\item{gdepth}{is an optional argument for the cell depth.  gdepth defaults to 10 m}

\item{database}{is an optional argument for export format.  Default is database (1).  Spreadsheet =0}

\item{sheet}{is an optional argument to denote which sheet of the DirNameFile
want to use. This defaults to the first sheet if not set.}

\item{SurveyName}{name of survey as in excel file}
}
\description{
function to export integration from variables in a set of EV files.
There are two excel files for inputs (paths, parameters to export).  The exports
go into a directory that is automatically named by date.  This directory is in the
export directory listed in the DirNameFile.  The EVfile is not saved in the process.
}
